fix(ptokens-request-processor): fix racePromise usage in callContractFunctionAndAwait()
The transaction object returned by ethers.js has a wait method that can be called
to await for the transaction receipt.
Passing the method to ptokens-utils racePromise utility in the following way:
    logic.racePromise(_txTimeout, _tx.wait, [])
does not work as expected as wait would lose its context (the transaction).
Thus, wait needs to be passed to racePromise by binding it to the transaction.